tests/wpt/tests/encoding-detection/he-ISO-8859-8.tentative.html

<!doctype html>
<title>he ISO-8859-8</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<p>.םיוות דודיק ןחבמ והז</p>
<script>
setup({explicit_done:true});
onload = function() {
    test(function() {
        assert_equals(document.characterSet, "ISO-8859-8", 'Expected ISO-8859-8');
    }, "Check detection result");
      done();
};
</script>
